.conda/requirements.txt
Flask
Flask-SQLAlchemy
requests
python-dotenv
google-api-python-client